requirements.txt
Cython==0.29.6
decorator==4.3.2
freetype-py==2.0.0.post6
future==0.17.1
glfw==1.7.1
glumpy==1.0.6
imageio==2.5.0
multipledispatch==0.6.0
networkx==2.2
nose==1.3.7
numpy==1.16.2
Pillow==5.4.1
pkg-resources==0.0.0
pygame==1.9.4
pyglet==1.4.0b1
pymesh2==0.2.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyrender==0.1.14
pyrr==0.10.1
scipy==1.2.1
six==1.12.0
triangle==20190115.1
trimesh==2.37.3
